shivaai_backend/requirements.txt
# Updated requirements for SHIVAAI v2.0.0
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# LLM & AI
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
google-generativeai>=0.8.3

# Embeddings & Vector Store
sentence-transformers>=3.0.0
torch>=2.0.0
transformers>=4.40.0
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.1.0
scikit-learn>=1.3.0
numpy>=1.26.0,<2.0.0

# NLP
nltk>=3.9.0
huggingface-hub>=0.20.0

# PDF & OCR
pdfplumber>=0.11.0
pytesseract>=0.3.10
pillow>=10.0.0

# Database
sqlalchemy>=2.0.0

# Streaming
sse-starlette>=2.0.0

# Utilities
python-multipart>=0.0.9
python-dotenv